[fix] 오타 수정, calendar 함수 옮기기
